0.0.040

+++ bot/callbacks/wordlie_callback.py
@@ -2,14 +2,36 @@ from aiogram import Router, F
 from aiogram.types import CallbackQuery
 from aiogram.fsm.context import FSMContext
 from gameControll.game import game
+import logging
 
 router = Router()
 @router.callback_query(F.data == "wordlie_from_friend")
 async def game_wordlie_from_friend(call: CallbackQuery, state: FSMContext):
-    await call.message.delete()
-    await state.update_data(state="f_in_game_wordlie")
-    await call.message.answer("Введите слово")
-    await state.set_state(game.state)
+    a:dict = await state.get_data()
+    if a == {} or not a["state"].startswith(("in_game", "f_in_game")):
+        await call.message.delete()
+        await state.update_data(state="f_in_game_wordlie")
+        await call.message.answer("Введите слово")
+        await state.set_state(game.state)
+        logging.info(
+                    f"пользователь принял вордли игру от друга",
+                        extra={"username": call.from_user.username,
+                        "state": await state.get_data(),
+                        "handler_name": "game_wordlie_from_friend",
+                        "params":"nothing"}
+                        )
+    else:
+        await call.message.answer("Закончи игру, прежде чем начинат новую")
+        logging.info(
+                    f"пользователь пытался начать игру вордли от друга, но у него уже есть начатая игра",
+                        extra={"username": call.from_user.username,
+                        "state": await state.get_data(),
+                        "handler_name": "game_wordlie_from_friend",
+                        "params":"nothing"}
+                        )
+    
+
+
 
 @router.callback_query(F.data == "wordlie_diss")
 async def disconnect(call: CallbackQuery):
@@ -20,3 +42,10 @@ async def disconnect(call: CallbackQuery):
         text=f"игрок @{call.from_user.username} отклонил игру"
     )
     del game.wordlie.rooms[call.from_user.id]
+    logging.info(
+                f"пользователь отклонил игру в вордли от друга",
+                    extra={"username": call.from_user.username,
+                    "state": "nothing",
+                    "handler_name": "disconnect",
+                    "params":"nothing"}
+                    )
